Extract DbDataExport class from DbExport
